pyproject.toml
[tool.poetry]
name = "gan-sr-wind-field"
version = "0.1.0"
description = ""
authors = ["jacobwulffwold <61497974+jacobwulffwold@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "gan_sr_wind_field"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
mayavi = "^4.8.1"
matplotlib = "^3.7.1"
netcdf4 = "^1.6.3"
tqdm = "^4.65.0"
scipy = "^1.10.1"
llvmlite = "^0.39.1"
scikit-learn = "^1.2.2"
black = "^23.1.0"
graphviz = "^0.20.1"
torchviz = "^0.0.2"
tensorboardx = "^2.6"
progressbar2 = "^4.2.0"
opencv-python = "^4.7.0.72"
torch-tb-profiler = "^0.4.1"
torchvision = "^0.15.1"
torchaudio = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
